contract-analyzer-webui/backend/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0
sentence-transformers==2.3.1
pydantic==2.5.3
numpy==1.26.3
pandas==2.1.4
tqdm==4.66.1
ollama==0.1.6
reportlab==4.0.7
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
